Add working auth for sandbox
